Sidebar: extract drag, tree, and navigation into their own hooks
